requirements.txt

#llm 
huggingface_hub
transformers 
datasets 
accelerate 
peft 
bitsandbytes 
torch 
tiktoken 
blobfile 
protobuf 
sentencepiece
flash-attn==1.0.5
trl 

#data
sqlalchemy==2.0.38
asyncpg==0.30.0
psycopg2-binary==2.9.10
pandas

#general
stockfish
python-dotenv
tqdm
chess 


#eval
matplotlib
numpy